messaging_app/requirements.txt
Django>=5.2
djangorestframework
django-cors-headers